Add filter times configuration
See #5

See #17

The user now has the option to automatically turn the filter on and
off. If they select the option, they can choose at what times the filter
is turned on and off. Later the user may have the choice to not select
the turn on and off times manually, but to automatically turn on and of
at sunrise and sunset.