Integrate LocationPreference with sun mode
See #29

The location is now updated silently when:
* The user selects the sun mode
* The filter is automatically turned on or off

Furthermore the location is explicitly updated when the user taps the
LocationPreference.

The AutomaticFilterChangeReceiver has been simplified, since it doesn't
have to find the sunset nor a location anymore, but just reads the
current on and off times. The on and off times are saved in the keys off
the new FilterTimePreferences. They change to sunrise and sunset times
when the sun mode is selected, automatically backing up the custom
times.

When the location fails to update when it has been explicitly called by
the user, a toast with a notice is displayed.

When the user selects the sun mode with a unset location, sensible
default sunrise and sunset values are set as turn on and off times.