Change state system
See #29

The user can now toggle pause/resume with the floating action button and
toggle disable/enable with the switch in the app bar.

When Red Moon is disabled, there is no notification and all settings are
grayed out.

When Red Moon is enabled and the filter is paused, the settings are
accessible and there is a notification in the notification tray, but the
filter is not running.

When Red Moon is enabled and the filter is resumed, the settings are
accessible and there is a notification in the notificaton tray and the
filter is running.

The filter will still be turned automatically when Red Moon is enabled.