Fix some issues with opening Activities from prefs
It seems that there where a few things going on. First Android couldn't
find the Activities, which is solved by this commit by declaring
specific actions for these Activities.

Furthermore, the call to actionBar.setDisplayHomeAsUpEnabled throws an
RunTimeError, it seems that method is abstract for that object.

Lastly, Eventbus doesn't allow classes to initialize as an event class
without subscribing to any events as was the case in
SecureSuspendFragment. To fix this, this commit adds one event
subscription to this class and does nothing with it.