Show correct animation when closing subactivity
When the TimeToggleActivity or SecureSuspendActivity where closed with
the arrow in the action bar, the wrong animation would display, making
it seem that the main activity was opened over the subactivity. Now the
close activity animation is displayed, so it looks like you are
returning to the main activity (which you are, the subactivities aren't
kept 'under' the main activity).

To get android to display this animation correctly I used a sort of a
workaround, overriding the functionality of the arrow in the action bar
with calling finish() on the activity. I'm not aware of any other way to
do this, so I think this should do fine.