Fix bug where filter would animate incorrectly
There was a bug where Red Moon would animate the filter in correctly,
going from a high intensity quickly to the normal filter, instead of
from transparent to the normal filter. This occurred when the main
activity was turning on the filter for the first time in its lifecycle
or when the activity wasn't running at all (stopped).

To replicate on older versions: Open the Red Moon activity and turn off
the server. Then close Red Moon completely (for example by sliding it
out of the multitasking menu on vanilla Android). Open Red Moon and use
the switch in the app bar to start the filter, now it should animate
incorrectly.

As far as I see this was caused by the animate functions in
ScreenFilterView always using the current filter level as the start of
the animation, which isn't initialized correctly on the start of the
presenter and is also effected by the PreviewState. This commit lets the
caller explicitly choose whether to animate from off or animate from the
current level.