Fix bug where the switch wasn't set correctly
I don't know if this bug occurs on any other devices, but on my Nexus
5X the switch in the app bar is set incorrectly after the following
steps:

1. Open Red Moon
2. Start the filter
3. Use the home button to exit Red Moon
4. Stop the filter through the notification
5. Reopen Red Moon through the launcher

After these steps the switch in the activity is still on for me,
although the filter is off.

To fix this, I've added an attempt to set the switch correctly on the
Resume event. Since the Switch may not be initialized (if the app is
opened after closing completely), we have to catch the
NotInitializedException. As far as I've found, there is no other way to
check if a lateinit variable is initialized in Kotlin.

This fix may be a bit crude, but I don't think it does any harm, and it
does fix the bug at least on my phone.