Add very crude proof of concept for root mode
See #150

It will request root after turning on the filter, which is quite
inconvenient, because you then have to turn of the filter through the
notification to grant the permission (a transparent filter is still
being drawn). It is also *very* slow so be patient when trying it. But
it seems to work (on my phone at least)!