Launch subscreens from code
Right now when you launch subscreens, the intent is caught by both the
debug version and release version, so if they're installed alongside one
another Android will ask which app should open. That's not desirable.
There might be a way to do it in xml but the documentation is bad and
I'm tired of wrestling with this bug.