Don't stop service if overlay permission is denied
Overlay permission is also denied if a permission dialog is open. So, we
don't want to kill the service, because then we can't resume filtering
when the permission dialog is closed. This is a bit hack-y right now
because the service will be started indefinitely if the user actually
has not granted overlay permission, but that's a much less user-facing
bug so we can put it off for now.